chore(legacy): move oidc utils to lib package
